Move Screen to runtime
