+++ circuit-runtime/src/androidMain/kotlin/com/slack/circuit/runtime/Screen.kt
@@ -1,0 +1,7 @@
+package com.slack.circuit.runtime
+
+import android.os.Parcelable
+import androidx.compose.runtime.Immutable
+
+@Immutable
+public actual interface Screen : Parcelable
