+++ circuit-runtime/src/jvmMain/kotlin/com/slack/circuit/runtime/Screen.kt
@@ -1,0 +1,6 @@
+package com.slack.circuit.runtime
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+public actual interface Screen
