--- circuit/src/androidMain/kotlin/com/slack/circuit/Screen.android.kt
@@ -1,8 +1,0 @@
-// Copyright (C) 2022 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit
-
-import android.os.Parcelable
-import androidx.compose.runtime.Immutable
-
-@Immutable public actual interface Screen : Parcelable
