+++ circuit/src/commonMain/kotlin/com/slack/circuit/Presenter.kt
@@ -4,6 +4,7 @@ package com.slack.circuit
 
 import androidx.compose.runtime.Composable
 import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.Screen
 import kotlinx.coroutines.flow.MutableSharedFlow
 
 /**
