+++ circuit/src/commonMain/kotlin/com/slack/circuit/Ui.kt
@@ -5,6 +5,7 @@ package com.slack.circuit
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.Screen
 
 /**
  * Represents a composable UI for the given [UiState]. Conventionally, this should just be the
