--- circuit/src/jvmMain/kotlin/com/slack/circuit/Screen.jvm.kt
@@ -1,7 +1,0 @@
-// Copyright (C) 2022 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit
-
-import androidx.compose.runtime.Immutable
-
-@Immutable public actual interface Screen
