+++ circuit/src/jvmTest/kotlin/com/slack/circuit/EventListenerTest.kt
@@ -11,6 +11,7 @@ import app.cash.molecule.launchMolecule
 import app.cash.turbine.Turbine
 import com.google.common.truth.Truth.assertThat
 import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.Screen
 import java.time.LocalTime
 import java.time.ZoneOffset.UTC
 import java.util.concurrent.TimeUnit
